refactor(seances): TeacherSeances sous 300 lignes — éclatement complet (H6)
Éclatement du god-component en composable + sous-composants, CSS VERBATIM :
- TeacherSeances.vue 769→295 (orchestrateur)
- useTeacherSeances.js (272) état + chargement (cache) + filtres + actions visio
- TeacherSeancesFilters.vue (167) filtres matière/statut/période (v-model + apply/reset)
- TeacherSeancesStats.vue (123) grille des 4 compteurs
Tests: montage (intact) + composable + sous-composants = 8 verts. vite build vert.
Dette pré-existante conservée à l'identique et documentée :
- @media .seance-header/.seance-info-grid/.action-info/.action-buttons/.btn-action :
  cible des éléments désormais internes à SeanceCard (extrait #28), inertes via le
  scoping — gardées telles quelles dans l'orchestrateur.
